Se trabaja en  el concepto de amigos ya que las personas pueden tener amigos pero las empresa pueden tener seguidores ya esta toda la logica dinamicamente con los resolvers
